实现 Chronon 能量网络系统核心功能
- IChrononNode：新增节点接口，定义网络 UUID、节点类型、产能/耗能/容量等行为，并实现节点放置时的自动网络合并逻辑（连接相邻网络/合并网络）
- ChrononNetwork：实现完整网络图结构，包括节点增删、网络合并与分裂检测、能量存储接口（IChrononStorage）、序列化/持久化支持
- ChrononNetworkData：实现网络数据的 SavedData 持久化存储与加载

+++ src/main/java/org/moshang/tempusetchaos/api/IChrononStorage.java
@@ -1,11 +1,11 @@
 package org.moshang.tempusetchaos.api;
 
 public interface IChrononStorage {
-    int receiveChronon(int amount, boolean simulate);
-    int extractChronon(int amount, boolean simulate);
+    long receiveChronon(long amount, boolean simulate);
+    long extractChronon(long amount, boolean simulate);
 
-    int getCapacity();
-    int getChrononStored();
+    long getCapacity();
+    long getChrononStored();
 
     default boolean canReceive() { return true; }
     default boolean canExtract() { return true; }
